+++ views/user/index.php
@@ -1,21 +1,19 @@
 <?php
 
 use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\widgets\ModalButton;
-use humhub\widgets\ModalDialog;
+use humhub\libs\Html;
 
 ?>
 
-<?php ModalDialog::begin(); ?>
+<?php $this->beginContent('@user/views/account/_userSettingsLayout.php') ?>
+
 <?php $form = ActiveForm::begin(); ?> 
-<div class="modal-body">
-    <?= $form->field($model, 'darkMode')->radioList($model->getOptions()); ?>
-	    <?= Yii::t('DarkModeModule.base', 'Choose "Follow system" to automatically switch between light and dark mode according to your browser or system preferences.')
-    ?>
-</div>
-<div class="modal-footer">
-    <?= ModalButton::cancel() ?>
-    <?= ModalButton::submitModal() ?>
-</div>
+    <div class="form-group">
+        <?= $this->render('form', ['model' => $model, 'form' => $form]); ?>
+    </div>
+    <div class="form-group">
+        <?= Html::submitButton(Yii::t('VerifiedModule.base', 'Save'), ['class' => 'btn btn-primary', 'data-ui-loader' => '']); ?>
+    </div>
 <?php ActiveForm::end(); ?>
-<?php ModalDialog::end(); ?>
+
+<?php $this->endContent(); ?>